Create json__commands__v6
Reads through a json file that is indexed with base(36) character system where the output of the previous function dictates which of the next 36 pathways it'll load. It also loads all functions on the fly, as needed, pre process. No functions not used in the "user click input" are loaded into memory, ever.