Create example__json__file__to_be_used_with_json__commands__v6
If you look closely you can see that in the function listings, it diverges into a few different pathways. This is dictated by the output of the previous function that was iterated from this json file. So the function "get_json_commands" will have a certain type of output during it's processing, that'll dictate the next index number, 0, 1, or 2, which when concantenated into a string with the previous "json indexes" creates the next pathway for the system to load, and therefore the next function it "includes" in the program. So  "get_json_commands" can go to: 0 = notify, 1 = get_header, or 2 = get_pasteinplace. These are functions that are stored on a php page and have only 2 parameters: $v object, and the $user object.